Adapt LibLog to MvvmCross
