+++ MvvmCross/Platform/Platform/Logging/IMvxLog.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace MvvmCross.Platform.Logging
+{
+    public interface IMvxLog
+    {
+        bool Log(MvxLogLevel logLevel, Func<string> messageFunc, Exception exception = null, params object[] formatParameters);
+    }
+
+    delegate bool Logger(MvxLogLevel logLevel, Func<string> messageFunc, Exception exception = null, params object[] formatParameters);
+}
