+++ MvvmCross/Platform/Platform/Logging/IMvxLogProvider.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace MvvmCross.Platform.Logging
+{
+    public interface IMvxLogProvider
+    {
+        IMvxLog GetLogFor<T>();
+
+        IMvxLog GetLogFor(string name);
+
+        IDisposable OpenNestedContext(string message);
+
+        IDisposable OpenMappedContext(string key, string value);
+    }
+}
