+++ MvvmCross/Platform/Platform/Logging/MvxLogLevel.cs
@@ -1,0 +1,15 @@
+﻿namespace MvvmCross.Platform.Logging
+{
+    /// <summary>
+    /// The log level.
+    /// </summary>
+    public enum MvxLogLevel
+    {
+        Trace,
+        Debug,
+        Info,
+        Warn,
+        Error,
+        Fatal
+    }
+}
