+++ MvvmCross/Platform/Platform/Platform/Logging/IMvxLog.cs
