--- MvvmCross/Platform/Platform/Platform/MvxLog.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using MvvmCross.Platform.Logging;
-
-namespace MvvmCross.Platform.Platform
-{
-    public class MvxLog : IMvxLog, ILog
-    {
-        public bool Log(LogLevel logLevel, Func<string> messageFunc, Exception exception = null, params object[] formatParameters)
-        {
-            Mvx.Trace(messageFunc.Invoke(), formatParameters);
-            return true;
-        }
-    }
-}
